Extract out materials and common modules
