+++ assets/js/main.js
@@ -3,16 +3,11 @@
 
 const $ = require('jquery');
 
-// initialise router (eg. run conditional code for certain URLs)
-const Grapnel = require('./libs/grapnel');
-const router = new Grapnel({ pushState: true });
-
-// initialise Vue
+/**
+ * Initialise Vue
+ */
 const Vue = require('../../node_modules/vue/dist/vue.common');
 Vue.options.delimiters = ['<%', '%>'];
-
-const analytics = require('./modules/analytics');
-const utils = require('./modules/utils');
 
 /**
  * Bootstraps
@@ -20,51 +15,25 @@ const utils = require('./modules/utils');
 const raven = require('./bootstraps/raven');
 raven.init(Vue);
 
-// grab main script element (for querying data attributes)
-const $thisScript = $('#js-script-main');
-
-// bind mobile nav show/hidew button
-$('#js-mobile-nav-toggle').on('click', function(e) {
-    e.preventDefault();
-    $html.toggleClass('show-off-canvas');
-});
-
-// look up a cookie's value
-const getCookieValue = a => {
-    let b = document.cookie.match('(^|;)\\s*' + a + '\\s*=\\s*([^;]+)');
-    return b ? b.pop() : '';
-};
-
-// toggle contrast mode (we do this in JS to avoid breaking caching)
-const $html = $('html');
-const isHighContrast = getCookieValue('contrastMode'); // @TODO get from config
-if (isHighContrast === 'high') {
-    $html.addClass('contrast--high');
-    $('#js-contrast-standard').show();
-    $('#js-contrast-high').hide();
-} else {
-    $('#js-contrast-standard').hide();
-    $('#js-contrast-high').show();
-}
-
-// show/hide overlay pane
-$('#js-close-overlay').on('click', () => {
-    $('#js-overlay').hide();
-});
-
 /**
  * Load modules
  */
+require('./modules/common').init();
 require('./modules/tabs').init();
 require('./modules/carousel').init();
 require('./modules/heroImages').init();
 require('./modules/logos').init();
+require('./modules/materials').init();
 
-// if the env is PRODUCTION then we only load analytics
-// if the page's domain name matches our live domain
-// setup google analytics
+/**
+ * Analytics
+ * If the we are in the live environment then load analytics
+ * @see main.njk for where _BLF.blockAnalytics is set
+ */
+const analytics = require('./modules/analytics');
 if (!_BLF.blockAnalytics) {
-    // set in main.njk
+    // grab main script element (for querying data attributes)
+    const $thisScript = $('#js-script-main');
 
     // get per-environment GA code
     const uaCode = $thisScript.data('ga-code');
@@ -128,84 +97,3 @@ if (!_BLF.blockAnalytics) {
         }
     });
 }
-
-// create Vue element for order form
-// we use a regex for this URL to allow welsh URLs to match too
-let fundingPagePath = /\/funding\/funding-guidance\/managing-your-funding\/ordering-free-materials/;
-router.get(fundingPagePath, () => {
-    let allOrderData = {};
-    let langParam = 'lang';
-    // make sure we only allow valid language options
-    let isValidLangParam = param => ['monolingual', 'bilingual'].indexOf(param) !== -1;
-
-    // save the language preference in the form so we can redirect to it
-    let storeLangPrefInForm = newState => $('#js-language-choice').val(newState);
-
-    new Vue({
-        el: '#js-vue',
-        data: {
-            orderData: allOrderData,
-            itemLanguage: null
-        },
-        created: function() {
-            // check for a ?lang param and show the relevant products (if valid)
-            let params = utils.parseQueryString();
-            if (params[langParam] && isValidLangParam(params[langParam])) {
-                storeLangPrefInForm(params[langParam]);
-                this.itemLanguage = params[langParam];
-            }
-        },
-        methods: {
-            // swap between languages for product list
-            toggleItemLanguage: function(newState) {
-                if (isValidLangParam(newState)) {
-                    this.itemLanguage = newState;
-                    storeLangPrefInForm(newState);
-                    // add this to the URL
-                    if (history.replaceState) {
-                        history.replaceState(null, null, `?${langParam}=${newState}`);
-                    }
-                }
-            },
-            // look up the quantity of a given item, defaulting to its
-            // value when the page was loaded (eg. from session cookie)
-            getQuantity: function(code, valueAtPageload) {
-                if (this.orderData[code]) {
-                    return this.orderData[code].quantity;
-                } else {
-                    return valueAtPageload;
-                }
-            },
-            // work out if the user has anything in their "basket"
-            // eg. should the data form be disabled or not
-            isEmpty: function() {
-                let quantity = 0;
-                for (let o in this.orderData) {
-                    quantity += this.orderData[o].quantity;
-                }
-                return quantity === 0;
-            },
-            // increment/decrement a product in the user's basket via AJAX
-            changeQuantity: function(e) {
-                let $elm = $(e.currentTarget);
-                const $form = $elm.parents('form');
-                const url = $form.attr('action');
-                let data = $form.serialize();
-                // is this an increase or decrease button?
-                data += '&action=' + $elm.val();
-                $.ajax({
-                    url: url,
-                    type: 'POST',
-                    data: data,
-                    dataType: 'json',
-                    success: response => {
-                        // update the basket data from the session
-                        allOrderData = response.allOrders;
-                        // this triggers a Vue update (and needs a babel plugin to work in IE)
-                        this.orderData = Object.assign({}, this.orderData, response.allOrders);
-                    }
-                });
-            }
-        }
-    });
-});
